Removed old implementation of gestures.
