+++ __tests__/gestures/spread.spec.js
@@ -1,5 +1,5 @@
-const Gesture = require("../../src/worker/Gesture-v2");
-const gestures = require("../../src/worker/gestures-v2");
+const Gesture = require("../../src/worker/Gesture");
+const gestures = require("../../src/worker/gestures");
 
 it("returns a Gesture instance", () => {
   const gesture = gestures.spread({ x: 100, y: 100, distance: 25 });
